refactor: Extract reporting capability to class
This way, we can introduce a MattermostReporter
